+++ src/app/layout.tsx
@@ -1,7 +1,8 @@
 import { Geist, Geist_Mono } from 'next/font/google';
 import { VIEWPORT, METADATA } from '@/configs';
 import { ThemeProvider } from '@/providers/ThemeProvider';
-import { LanguageProvider } from '@/providers/LanguageProvider'
+import { LanguageProvider } from '@/providers/LanguageProvider';
+import Web3Provider from '@/providers/Web3Provider';
 import './globals.css';
 
 const geistSans = Geist({
@@ -23,9 +24,11 @@ const RootLayout = ({children}: {children: React.ReactNode}) => {
             <body>
                 <ThemeProvider>
                     <LanguageProvider />
-                    <main>
-                        {children}
-                    </main>
+                        <Web3Provider>
+                            <main>
+                                {children}
+                            </main>
+                        </Web3Provider>
                 </ThemeProvider>
             </body>
         </html>